pinocchio/interface: TokenInstruction: 45/UnwrapLamports in From<u8>
`TokenInstruction` has an `UnwrapLamports` variant that should probably
be part of the `From<u8>` conversion as well?

SPL Token2022 does not have this instruction.  Nor is it present in the
SPL Token.  So, maybe it should have a different value?  Perhaps
something closer to the end of the u8 range?