Cargo.toml: Move all dependency versions to workspace level
All workspace creates already share a single workspace level Cargo.lock
file.  And most crates already have some of the dependencies specified
as the workspace level dependencies.  It seems more consistent to use
same dependency version for everything in the workspace.  It also makes
it easier to manage dependency versions in a consistent manner, in
particular when a third party dependency version is updated.