pinocchio/interface: TokenInstruction,AuthorityType: Add `Copy` and `Eq`
There seems to be no reason not to implement either `Copy` or `Eq` for
these enumeration types.

They are expected to take a single byte in the future.  So the by-value
semantics from `Copy` is natural.

Considering they already implement `PartialEq`, adding `Eq` is obviously
correct, as no values are expected to be incomparable.